app/requirements.txt
flask
flask-sock
psycopg2-binary
kafka-python
